Remove `Page` from the controller names
This was an action item discussed at part of the team Techtros.
