+++ app/views/declaration/carrier_details.scala.html
@@ -28,7 +28,7 @@
     title = messages("supplementary.carrier.title"),
     appConfig = appConfig
 ) {
-    @helper.form(CarrierDetailsPageController.saveAddress(), 'autoComplete -> "off") {
+    @helper.form(CarrierDetailsController.saveAddress(), 'autoComplete -> "off") {
         @helper.CSRF.formField
 
         @components.back_link("/customs-declare-exports/declaration/representative-details")
