+++ app/views/declaration/consignee_details.scala.html
@@ -28,7 +28,7 @@
     title = messages("supplementary.consignee.title"),
     appConfig = appConfig
 ) {
-    @helper.form(ConsigneeDetailsPageController.saveAddress(), 'autoComplete -> "off") {
+    @helper.form(ConsigneeDetailsController.saveAddress(), 'autoComplete -> "off") {
         @helper.CSRF.formField
 
         @components.back_link("/customs-declare-exports/declaration/exporter-details")
