+++ app/views/declaration/declarant_details.scala.html
@@ -30,7 +30,7 @@
     title = messages("supplementary.declarant.title"),
     appConfig = appConfig
 ) {
-    @helper.form(DeclarantDetailsPageController.saveAddress(), 'autoComplete -> "off") {
+    @helper.form(DeclarantDetailsController.saveAddress(), 'autoComplete -> "off") {
         @helper.CSRF.formField
 
         @components.back_link("/customs-declare-exports/declaration/consignee-details")
