+++ app/views/declaration/exporter_details.scala.html
@@ -28,7 +28,7 @@
     title = messages("supplementary.consignor.title"),
     appConfig = appConfig
 ) {
-    @helper.form(ExporterDetailsPageController.saveAddress(), 'autoComplete -> "off") {
+    @helper.form(ExporterDetailsController.saveAddress(), 'autoComplete -> "off") {
         @helper.CSRF.formField
 
         @components.back_link("/customs-declare-exports/declaration/consignment-references")
