+++ app/views/declaration/items_summary.scala.html
@@ -73,7 +73,7 @@
                         </ol>
                     </td>
                     <td scope="row">
-                        <a id="item_@index--change" href="@routes.ProcedureCodesPageController.displayPage(item.id)">@messages("site.change")</a>
+                        <a id="item_@index--change" href="@routes.ProcedureCodesController.displayPage(item.id)">@messages("site.change")</a>
                     </td>
                     <td scope="row">
                         <a id="item_@index--remove" class="button button--gray" href="@routes.ItemsSummaryController.removeItem(item.id)">@messages("site.remove")</a>
