+++ app/views/declaration/package_information.scala.html
@@ -40,7 +40,7 @@
     appConfig = appConfig
 ) {
 
-    @components.back_link(routes.ItemTypePageController.displayPage(itemId).url)
+    @components.back_link(routes.ItemTypeController.displayPage(itemId).url)
 
     @components.error_summary(form.errors)
