+++ app/views/declaration/representative_details.scala.html
@@ -33,7 +33,7 @@
 
     @components.back_link("/customs-declare-exports/declaration/declarant-details")
 
-    @helper.form(RepresentativeDetailsPageController.submitRepresentativeDetails(), 'autoComplete -> "off") {
+    @helper.form(RepresentativeDetailsController.submitRepresentativeDetails(), 'autoComplete -> "off") {
         @helper.CSRF.formField
 
         @components.error_summary(form.errors)
